Structural overhaul. Server code moved into a 'server' directory, express routes split up into multiple files in 'routes' dir, api methods moved into its own file
